requirements-docker.txt
# Core dependencies for Docker container
fastapi
uvicorn[standard]
pydantic
aiosqlite
click

# Security: Pin minimum secure versions
h11>=0.16.0
jinja2>=3.1.6
requests>=2.32.4
setuptools>=78.1.1

# MCP SDK for refresh functionality
mcp

# Synvya SDK for Nostr functionality
synvya-sdk

# Security and rate limiting

# Environment and configuration management
python-dotenv

# HTTP client for health checks and database service communication
httpx
aiohttp

# Note: safety excluded for Docker container due to pydantic version conflicts
# Use 'pip install safety' manually for local development security audits 